Allow mutexes to be released out of order
Some algorithms which use mutexes for locking require that the mutexes
be released out of order. The AcpiInitializeObjects function from
ACPICA ends up using such an algorithm.

Previously, mutex_release would not allow mutexes to be released
out-of-order and would instead cause a system crash. Due to the above,
this was deemed to be too restrictive.