Fix broken strncpy implementation
The original implementation of strncpy in lib/string.c was incorrect.
Specifically, the code to fill the remainder of the buffer with zeroes
after copying the string did not work correctly if the source was too
long. In that case, n would be 0xffffffff when entering the second
loop, causing the buggy implementation to try to write 0's over the
entire address space.