Fix ambiguous error codes from gprintf
In the event of some sort of failure, the previous implementation of
gprintf would return a positive error code. However, this was
indistinguishable from a successful result.

In order to ensure that the return value from gprintf can be checked
for errors, and to conform to the standards used by implementations of
other printf-style functions in the C standard, the error codes are now
turned negative before being returned from gprintf.