Make copies of important multiboot info on boot
Previously, the multiboot structure was used directly whenever anything
from it was required. However, the multiboot information structures can
theoretically be placed anywhere in memory on boot; specifically, GRUB
will end up placing them in lower memory (<1MiB).

This could become an issue if this memory needs to be reclaimed during
the boot process, since there would be no easy way of finding the
memory used by these structures and marking it as in use.

In order to rectify this issue, relevant multiboot structures are now
copied in a more useful format immediately after the early memory
manager is initialized. This allows memory used by those structures to
be reclaimed immediately after the full memory manager is initialized.