Reserve space for kmalloc_early
Previously, kmalloc_early would simply use available space after the
end of the kernel binary. However, it is possible for the bootloader to
have loaded important structures there which should not be overwritten
yet.

In order to ensure these structures are not overwritten by dynamically
allocated memory, kmalloc_early now uses a dedicated section of memory
when performing its allocations. The size of this section is restricted
to 256KiB. Since the bootloader will know that this area is reserved
for use by the kernel, it should not place any data structures there.
This ensures that bootloader-allocated structures do not get stomped
before they are copied.