Allow frames under 1MiB to be marked as free
Previously, all frames before the end of the kernel would always be
marked as "KERNEL RESERVED". However, for some uses (e.g. initializing
SMP), it's important to be able to allocate frames <1MiB.

In order to accomodate this, the way that kernel reserved frames are
found has been changed. This will now allow frames in low memory to be
added to the free frame stack.