Improve unblocking serial port support
Previously, the support for unblocking I/O on serial ports was lacking.
In order to be able to implement a GDB stub using a serial port, it
will be necessary to such unblocking I/O to be possible.

Firstly, it is now possible to set a sink for received serial data to
be handled by. This handler is executed in an interrupt context, which
will allow debugging code that listens to the serial port to interrupt
existing code and temporarily suspend the system.

Secondly, two new functions have been added for performing serial I/O
from an interrupt context, where blocking is not possible. The
serial_send_spin and serial_receive_spin functions will simply spin
while waiting, rather than blocking the current thread. However, this
makes them unsuited to regular use. As such, their use should be
limited to debugging purposes only.