Implement setjmp and longjmp
In order to allow for error handling in certain situations, it is
necessary to have a mechanism for non-local jumps, which in C means the
setjmp and longjmp functions.

In addition, a specialized longjmp_interrupt function has been added.
This function allows the caller to specify that the longjmp should be
performed when returning from an interrupt.

Due to the fact that returning from a fault will retry the faulting
instruction, modification of the fault's return address is the only
viable way for a function to handle faults that it might cause during
execution. By using setjmp and putting the jmp_buf somewhere that the
fault handler will be able to find it, longjmp_interrupt can be used to
allow the faulting function to handle its own exceptions.