Fix possible regs32 corruption if r->esp is changed
By changing r->esp in an interrupt handler, it was possible to get code
in interrupt_common to move the regs32 struct incorrectly before
returning. The code for moving the regs32 struct always assumed that it
would be moving the structure to a higher memory address. By changing
r->esp to a lower address such that the destination of the move
overlaps the original, the code would corrupt the structure by moving
it in the wrong order.

In order to correct this, the actual moving of the regs32 structure has
been delegated to a call to the memmove function, which always performs
the copy in the correct order.