Allow a temporary page fault handler to be set
With the new ability to use setjmp/longjmp to call error handlers from
interrupt service routines, it is now possible to use this mechanism to
allow functions to handle their own page faults.

In order to allow this, kmem_page_set_temporary_fault_handler and
kmem_page_clear_temporary_fault_handler have been added, which allow a
function to register a longjmp to be executed if a page fault is
encountered.