Add kmem_page_preinit
Previously, the page fault handler was only registered when
kmem_page_init was called. However, since it may be necessary to make
use of memcpy_safe before this can happen, initialization of the page
fault handler has been moved to a separate kmem_page_preinit function.

This function can be called at any time and doing so will guarantee
that page fault handling is initialized and memcpy_safe can be called
safely. If it is not called before kmem_page_init is called, the page
fault handler will be initialized then.