Add a basic GDB stub
Previously, debugging had to be done using a GDB stub built into the
emulator to be used for testing. However, there were several minor
annoyances with doing this.

Primarily, debugging after an exception was difficult. If an exception
caused a crash, using the emulator's GDB stub would result in debugging
the interrupt handler itself, which was not of interest. By using a
custom GDB stub built into the kernel, we can ensure that debugging
after a crash initiated by an interrupt handler will debug the faulting
code, not the fault handler itself.

Unfortunately, the use of this GDB stub has several drawbacks. For
instance, debugging of hangs with interrupts disabled is not easy. Due
to the fact that interrupts are disabled, there is no way to enter the
GDB stub without single-stepping before the hang occurs.