Move to CMake for building
Previously, custom Makefiles were being used to build the kernel. This
was causing a number of annoyances; for instance, plain Makefiles do not
have any way to allow for the builder to specify extra options for
compilation short of either modifying the Makefile or setting
environment variables. The environment variables solution was not very
robust, and any modifications to Makefiles would need to be excluded
from commits to avoid any conflicts.

In order to address this, the build system has been moved to use CMake.
Not only is this generally more pleasant to use, it solves the above
issue by allowing for options to be set at build time. In addition to
this, it is now expected that builds be performed in a build
subdirectory to allow for cleaner separation between build artifacts and
code.