Change doxygen configuration
The previous configuration given to Doxygen has had some minor tweaks
applied to it. Firstly, the output is now optimized for C in that it
uses C terminology rather than generic terminology. Additionally,
undocumented members and structures are now hidden, since they are not
generally intended for use outside of the kernel itself. Lastly, static
inline functions defined in the header files are now properly included
in the output.