Fix broken mutex_try_acquire implementation
Due to a faulty implementation of mutex_try_acquire, mutexes acquired
using this function would not be fully acquired: they would be marked as
taken, but the owning thread of the mutex was not being set. Because of
this, when mutex_release was called on the previously acquired mutex,
that function would believe the mutex was being released by the wrong
thread and the system would crash.