Prevent unchecked_unwind from raising page faults
Previously, a corrupt stack could cause the unchecked_unwind function
to raise page faults. This would effectively obscure the cause of the
problem, since the page fault error's details would be displayed instead
of the details of the actual crash. By making unchecked_unwind handle
page faults, the attempt to generate a stack trace will stop when a
corrupt stack frame is encountered.