Implement support for a basic initrd filesystem
Previously, there was support for an initrd block device, but no
filesystems were supported to load from it. Support for an extremely
basic filesystem called the Sage Aasvogel Initialization Filesystem, or
SAIF, has been added to allow information to be read from the initrd.
This filesystem is read-only, and a special packer is run during the
build process to create this filesystem beforehand. Since SAIF is only
intended for use during system initialization, this should not be an
issue.

Additionally, a good chunk of the API for the VFS module has changed.
Unfortunately, a large chunk of the documentation for this module has
been out-of-date for a while due to the API being very much still in
development; in order to prevent confusion, this documentation has been
temporarily removed. New documentation will be written in the future
when the API becomes more stable.