Fix boot modules being overwritten in memory
During initialization, it is very important that any boot modules loaded
by the bootloader are kept in memory. Due to a flaw in the way that
kmem_phys_init was implemented, physical memory used to store these boot
modules was being marked as free; this could lead to boot modules being
overwritten by other data if there was sufficient demand for physical
memory.

This has now been corrected by created reserved memory regions for each
of the boot modules reported by the bootloader.