Fix load_kernel_symtab allocating wrong amount of space
Due to a typo, the load_kernel_symtab function was allocating space
based on the size of elf32_sym structs instead of the kernel_symbol
structs that it should have been allocating space for. Due to good luck,
these structures happened to have the same size, but this could have
become an issue if the kernel_symbol structure was modified.