Implement KERNEL_GDB_WAIT build option
Previously, a build option called KERNEL_GDB_WAIT was declared, but
never used. This option was meant to signal to the kernel that the
kernel GDB stub should stop execution as soon as it's initialized. This
option has now been implemented, and will modify the multiboot command
line to properly signal this intention to the kernel.