Include git revision in the version suffix
Previously, only the version number itself would ever be printed by a
built version of Sage Aasvogel. However, due to the early nature of the
project, this makes it extremely difficult to identify what version a
particular build is actually running.

To rectify this, the git commit hash is used as part of a suffix to the
version number; there is an option to disable this in the build process,
should a release ever be made.