Move definitions between CMakeLists.txt files
Previously, a good number of compiler flags and other such important
declarations were being made only in the src/kernel/CMakeLists.txt file.
However, in the future, it will be likely that there will be other
modules that require the same definitions. In order to allow this, those
definitions have been moved to src/CMakeLists.txt.