Report max packet size from GDB stub
When GDB connects to a remote stub, it will send a qSupported packet
specifying what features it supports and the stub should respond with
the featureset it supports. One of the features that can be reported
this way is the maximum packet size the stub is willing to accept. By
doing this, we prevent GDB from having to guess what this value is.